spark: Move pyoskar py files to root
