+++ oskar-spark/src/main/python/pyoskar/analysis.py
