+++ oskar-spark/src/main/python/pyoskar/sql.py
