+++ oskar-spark/src/test/python/pyoskar_tests/test_transformers.py
@@ -1,8 +1,7 @@
 import unittest
 
-from pyoskar.spark.sql import *
 from pyoskar_tests.test_utils import *
-from pyspark.sql.functions import *
+from pyspark.sql.functions import col
 
 
 LIMIT = 3
