+++ oskar-spark/src/test/python/pyoskar_tests/test_udfs.py
@@ -1,9 +1,7 @@
 import unittest
 
-from pyoskar.spark.sql import *
+from pyoskar.sql import *
 from pyoskar_tests.test_utils import *
-from pyspark.sql.functions import *
-
 
 LIMIT = 3
